BackEnd/analytics-service/requirements.txt
Django==5.0.14
django-cors-headers==4.7.0
pytz==2025.2
python-dotenv==1.1.1
PyMySQL==1.1.1
certifi==2025.8.3
reportlab==4.4.4